chore: move makeKVStoreFromMap
