+++ packages/cosmic-swingset/src/launch-chain.js
@@ -55,7 +55,8 @@ import { computronCounter } from './computron-counter.js';
  * @import {BlockInfo} from '@agoric/internal/src/chain-utils.js';
  * @import {SwingStoreKernelStorage} from '@agoric/swing-store';
  * @import {Mailbox, RunPolicy, SwingSetConfig} from '@agoric/swingset-vat';
- * @import {KVStore, BufferedKVStore} from './helpers/bufferedStorage.js';
+ * @import {BufferedKVStore} from './helpers/bufferedStorage.js';
+ * @import {KVStore} from '@agoric/internal/src/kv-store.js';
  * @import {ConfigProposal} from '@agoric/deploy-script-support/src/extract-proposal.js';
  * @import {ERef} from '@endo/far';
  * @import {QueueStorage} from './helpers/make-queue.js';
