+++ packages/swing-store/src/exporter.js
@@ -2,8 +2,7 @@ import sqlite3 from 'better-sqlite3';
 
 import { Fail, q } from '@endo/errors';
 
-import { dbFileInDirectory } from './util.js';
-import { getKeyType } from './util.js';
+import { dbFileInDirectory, getKVStoreKeyType } from './util.js';
 import { makeBundleStore } from './bundleStore.js';
 import { makeSnapStore } from './snapStore.js';
 import { makeSnapStoreIO } from './snapStoreIO.js';
@@ -133,7 +132,7 @@ export function makeSwingStoreExporter(dirPath, options = {}) {
    * section
    */
   function getHostKV(key) {
-    getKeyType(key) === 'host' || Fail`getHostKV requires host keys`;
+    getKVStoreKeyType(key) === 'host' || Fail`getHostKV requires host keys`;
     // @ts-expect-error unknown
     return sqlKVGet.get(key);
   }
@@ -151,7 +150,7 @@ export function makeSwingStoreExporter(dirPath, options = {}) {
    */
   async function* getExportData() {
     for (const { key, value } of sqlGetAllKVData.iterate()) {
-      if (getKeyType(key) === 'consensus') {
+      if (getKVStoreKeyType(key) === 'consensus') {
         yield [`kv.${key}`, value];
       }
     }
